Feat(Lotto): 중복된 로또 번호 예외 처리
- 중복된 로또 번호가 있을 경우 예외를 던짐

+++ src/main/java/lotto/LottoNumberValidator.java
@@ -9,7 +9,7 @@ public class LottoNumberValidator {
 
 
     public static boolean existDuplicate(List<Integer> numbers) {
-        long numbersTypeCount = numbers
+        int numbersTypeCount = (int)numbers
                 .stream()
                 .distinct()
                 .count();
